// src/api/backend/model/feedback.ts
/**
 * web-backend-swagger
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Category } from './category';

export interface Feedback { 
    categories: Array<Category>;
    taskId?: number;
    id: number;
    userId: number;
    lastUpdated: string;
    filled: boolean;
}